requirements.txt
# Core
python>=3.10
numpy>=1.24.0
pandas>=2.0.0

# STAC
pystac-client>=0.7.0
odc-stac>=0.3.0

# Geospatial
geopandas>=0.14.0
shapely>=2.0.0
pyproj>=3.6.0
rasterio>=1.3.0
xarray>=2023.1.0
rioxarray>=0.15.0

# CV/ML
raster-vision>=0.21.0
torch>=2.0.0
torchvision>=0.15.0

# Data
pyarrow>=14.0.0
fastparquet>=2023.10.0

# Orchestration
prefect>=2.14.0

# Dask
dask>=2023.12.0
distributed>=2023.12.0

# Utilities
click>=8.1.0
pydantic>=2.5.0
python-dateutil>=2.8.0
tqdm>=4.66.0

# Optional: visualization
matplotlib>=3.8.0
folium>=0.15.0

# Optional: backtesting (later stage)
# vectorbt
# backtrader
